Refactor Python builds to fit new skeleton
We now use pyproject.toml as much as possible. However, building the C
extension still requires setup.py
